Refactor codes
